wos creds
